+++ qlab/problems/__init__.py
@@ -8,6 +8,7 @@ from qlab.problems import (  # noqa: F401  (import = registration)
     grover,
     maxcut,
     noise,
+    qec_repetition,
     qft,
     qml_classifier,
     qpe,
@@ -18,4 +19,4 @@ from qlab.problems import (  # noqa: F401  (import = registration)
 )
 
 __all__ = ["state_prep", "maxcut", "bernstein_vazirani", "deutsch_jozsa", "simon", "grover", "qft", "qpe",
-           "shor", "vqe", "qml_classifier", "noise"]
+           "shor", "vqe", "qml_classifier", "noise", "qec_repetition"]